+++ .env.template
@@ -8,15 +8,8 @@ FLASK_DEBUG=true
 # Required to use Flask's session feature (for storing todos in a browser cookie in the first exercise): https://flask.palletsprojects.com/en/2.3.x/config/#SECRET_KEY
 SECRET_KEY=secret-key
 
+# Add your mongoDB info bellow
 
-# Add your trello info bellow
+CONNECTION_STRING=<Fill in>
 
-TRELLO_API_KEY=<Fill in>
-
-TRELLO_API_TOKEN=<Fill in>
-
-TRELLO_TODO_LIST_ID=<Fill in>
-
-TRELLO_DONE_LIST_ID=<Fill in>
-
-TRELLO_BOARD_ID=<Fill in>
+DATABASE_NAME=<Fill in>
