+++ .env.test
@@ -8,13 +8,4 @@ FLASK_DEBUG=true
 # Required to use Flask's session feature (for storing todos in a browser cookie in the first exercise): https://flask.palletsprojects.com/en/2.3.x/config/#SECRET_KEY
 SECRET_KEY=secret-key
 
-
-TRELLO_API_KEY=FakeAPIKey
-
-TRELLO_API_TOKEN=FakeAPIToken
-
-TRELLO_TODO_LIST_ID=fake-todo-list
-
-TRELLO_DONE_LIST_ID=fake-done-list
-
-TRELLO_BOARD_ID=fake-board-id
+CONNECTION_STRING=mongodb://fakemongo.com
