+++ pyproject.toml
@@ -10,6 +10,8 @@ Flask = "^2.2.5"
 python-dotenv = "^0.14.0"
 requests = "^2.31.0"
 pytest = "^8.0.0"
+pymongo = "^4.8.0"
+mongomock = "^4.2.0.post1"
 
 [tool.poetry.dev-dependencies]
