+++ todo_app/data/item.py
@@ -6,7 +6,3 @@ class Item:
 
     def __eq__(self, other): 
         return self.id == other.id and self.name == other.name and self.status == other.status
-
-    @classmethod
-    def from_trello_card(cls, card, list):
-        return Item(card['id'], card['name'], list['name'])
