--- todo_app/data/trello_items.py
@@ -1,61 +1,0 @@
-import requests
-import os
-from todo_app.data.item import Item
-
-def add_item(title):
-    reqUrl = "https://api.trello.com/1/cards"
-
-    query_params = {
-        "key" : os.getenv("TRELLO_API_KEY"),
-        "token" : os.getenv("TRELLO_API_TOKEN"),
-        "idList" : os.getenv("TRELLO_TODO_LIST_ID"),
-        "name" : title
-    }
-
-    response = requests.post(reqUrl, params=query_params)
-
-    response.raise_for_status()
-
-    print(response.text)
-
-
-def get_items():
-    board_id = os.getenv("TRELLO_BOARD_ID")
-
-    reqUrl = f"https://api.trello.com/1/boards/{board_id}/lists"
-
-    query_params = {
-        "key" : os.getenv("TRELLO_API_KEY"),
-        "token" : os.getenv("TRELLO_API_TOKEN"),
-        "cards" : "open"
-    }
-
-    response = requests.get(reqUrl, params=query_params)
-
-    response.raise_for_status()
-
-    response_list = response.json()
-
-    items = []
-
-    for trello_list in response_list:
-        for trello_card in trello_list['cards']:
-            item = Item.from_trello_card(trello_card, trello_list)
-            items.append(item)
-
-    return items
-
-
-def move_item_to_done(item_id):
-
-    reqUrl = f"https://api.trello.com/1/cards/{item_id}"
-
-    query_params = {
-        "key" : os.getenv("TRELLO_API_KEY"),
-        "token" : os.getenv("TRELLO_API_TOKEN"),
-        "idList" : os.getenv("TRELLO_DONE_LIST_ID")
-    }
-
-    response = requests.put(reqUrl, params=query_params)
-
-    response.raise_for_status()
