--- todo_app/tests/test_api_calls.py
@@ -1,48 +1,0 @@
-import os
-import pytest
-import requests
-from dotenv import load_dotenv, find_dotenv
-from todo_app import app
-
-@pytest.fixture
-def client():
-    file_path = find_dotenv('.env.test')
-    load_dotenv(file_path, override=True)
-
-    test_app = app.create_app()
-
-    with test_app.test_client() as client:
-        yield client
-
-
-class StubResponse():
-    def __init__(self, fake_response_data):
-        self.fake_response_data = fake_response_data
-
-    def json(self):
-        return self.fake_response_data
-    
-    def raise_for_status(self):
-        return 'Error'
-
-    
-def stub(url, params={}):
-    test_board_id = os.environ.get('TRELLO_BOARD_ID')
-
-    if url == f'https://api.trello.com/1/boards/{test_board_id}/lists':
-        fake_response_data = [{
-            'id' : '123abc',
-            'name' : 'To Do',
-            'cards' : [{'id' : '456', 'name' : 'Test card'}]
-        }]
-        return StubResponse(fake_response_data)
-    
-    raise Exception(f'Intergration test did not expect URL "{url}')
-
-
-def test_index_page(monkeypatch, client):
-    monkeypatch.setattr(requests, 'get', stub)
-    response = client.get('/')
-
-    assert response.status_code == 200
-    assert 'Test card' in response.data.decode()
